requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-generativeai==0.8.3
mcp==1.1.0
python-dotenv==1.0.1
pydantic==2.10.3
httpx==0.28.1
typing-extensions==4.12.2
pytest==8.3.4
pytest-asyncio==0.25.0
redis==5.2.1
# Database dependencies
sqlalchemy>=2.0.36
alembic>=1.13.3
asyncpg>=0.29.0
psycopg2>=2.9.9
# TODO: Replace with published package when available
# pyairtable-common==1.0.0